+++ vignettes/why-moderndive.R
@@ -1,4 +1,4 @@
-## ---- include = FALSE----------------------------------------------------
+## ---- include = FALSE---------------------------------------------------------
 knitr::opts_chunk$set(
   collapse = TRUE,
   comment = "#>"
@@ -14,11 +14,12 @@ library(ggplot2)
 library(moderndive)
 library(knitr)
 library(broom)
+library(patchwork)
 
 # https://www.youtube.com/watch?v=xjJ7FheCkCU
 set.seed(76)
 
-## ---- echo=FALSE---------------------------------------------------------
+## ---- echo=FALSE--------------------------------------------------------------
 library(moderndive)
 library(knitr)
 evals %>% 
@@ -26,23 +27,23 @@ evals %>%
   sample_n(5) %>% 
   kable()
 
-## ------------------------------------------------------------------------
+## -----------------------------------------------------------------------------
 library(moderndive)
 score_model <- lm(score ~ age, data = evals)
 
-## ------------------------------------------------------------------------
+## -----------------------------------------------------------------------------
 summary(score_model)
 
-## ------------------------------------------------------------------------
+## -----------------------------------------------------------------------------
 get_regression_table(score_model)
 
-## ------------------------------------------------------------------------
+## -----------------------------------------------------------------------------
 get_regression_points(score_model)
 
-## ------------------------------------------------------------------------
+## -----------------------------------------------------------------------------
 get_regression_summaries(score_model)
 
-## ---- echo=TRUE, eval=TRUE, fig.width=16/2.2, fig.height=9/2.2-----------
+## ---- echo=TRUE, eval=TRUE, fig.width=16/2.2, fig.height=9/2.2----------------
 library(ggplot2)
 ggplot(evals, aes(x = age, y = score, color = ethnicity)) +
   geom_point() +
@@ -50,7 +51,7 @@ ggplot(evals, aes(x = age, y = score, color = ethnicity)) +
        title = "Interaction model") +
   geom_smooth(method = "lm", se = FALSE)
 
-## ---- echo=TRUE, eval=TRUE, fig.width=16/2.2, fig.height=9/2.2-----------
+## ---- echo=TRUE, eval=TRUE, fig.width=16/2.2, fig.height=9/2.2----------------
 library(ggplot2)
 library(moderndive)
 ggplot(evals, aes(x = age, y = score, color = ethnicity)) +
@@ -59,43 +60,43 @@ ggplot(evals, aes(x = age, y = score, color = ethnicity)) +
        title = "Parallel slopes model") + 
   geom_parallel_slopes(se = FALSE)
 
-## ------------------------------------------------------------------------
+## -----------------------------------------------------------------------------
 get_regression_table(score_model)
 
-## ------------------------------------------------------------------------
+## -----------------------------------------------------------------------------
 get_regression_table(score_model) %>% 
   knitr::kable()
 
-## ------------------------------------------------------------------------
+## -----------------------------------------------------------------------------
 sqrt(diag(vcov(score_model)))
 
-## ------------------------------------------------------------------------
+## -----------------------------------------------------------------------------
 get_regression_table(score_model) %>% 
   pull(std_error)
 
-## ------------------------------------------------------------------------
+## -----------------------------------------------------------------------------
 get_regression_table(score_model)$std_error
 
-## ---- eval=FALSE---------------------------------------------------------
+## ---- eval=FALSE--------------------------------------------------------------
 #  fitted(score_model)
 
-## ---- echo=FALSE---------------------------------------------------------
+## ---- echo=FALSE--------------------------------------------------------------
 fitted(score_model)[1:10]
 
-## ---- eval=FALSE---------------------------------------------------------
+## ---- eval=FALSE--------------------------------------------------------------
 #  residuals(score_model)
 
-## ---- echo=FALSE---------------------------------------------------------
+## ---- echo=FALSE--------------------------------------------------------------
 residuals(score_model)[1:10]
 
-## ---- eval=FALSE---------------------------------------------------------
+## ---- eval=FALSE--------------------------------------------------------------
 #  get_regression_points(score_model)
 
-## ---- echo=FALSE---------------------------------------------------------
+## ---- echo=FALSE--------------------------------------------------------------
 get_regression_points(score_model) %>% 
   slice(1:10)
 
-## ---- echo=TRUE, eval=TRUE, fig.width=16/2.2, fig.height=9/2.2-----------
+## ---- echo=TRUE, eval=TRUE, fig.width=16/2.2, fig.height=9/2.2----------------
 score_model_points <- get_regression_points(score_model)
 
 # Histogram of residuals:
@@ -108,7 +109,7 @@ ggplot(score_model_points, aes(x = age, y = residual)) +
   geom_point() +
   labs(title = "Residuals over age")
 
-## ------------------------------------------------------------------------
+## -----------------------------------------------------------------------------
 new_evals <- evals %>% 
   sample_n(4) %>% 
   select(-score)
@@ -116,10 +117,10 @@ new_evals
 
 get_regression_points(score_model, newdata = new_evals)
 
-## ---- out.width = "500px", echo=FALSE, fig.align='center'----------------
+## ---- out.width = "500px", echo=FALSE, fig.align='center'---------------------
 knitr::include_graphics("kaggle.png")
 
-## ---- eval=FALSE---------------------------------------------------------
+## ---- eval=FALSE--------------------------------------------------------------
 #  library(tidyverse)
 #  library(moderndive)
 #  
@@ -135,51 +136,46 @@ knitr::include_graphics("kaggle.png")
 #    select(Id, SalePrice = SalePrice_hat)
 #  write_csv(submission, "submission.csv")
 
-## ---- out.width = "100%", echo=FALSE-------------------------------------
+## ---- out.width = "100%", echo=FALSE------------------------------------------
 knitr::include_graphics("leaderboard_orig.png")
 
-## ------------------------------------------------------------------------
+## -----------------------------------------------------------------------------
 get_regression_summaries(score_model)
 
-## ---- echo=FALSE, eval = FALSE-------------------------------------------
-#  # eval set to FALSE b/c patchwork does not exist on CRAN
-#  library(patchwork)
-#  p1 <- ggplot(evals, aes(x = age, y = score, color = ethnicity)) +
-#    geom_point() +
-#    labs(x = "Instructor age", y = "Teaching score", color = "Instructor\nEthnicity",
-#         title = "Interaction model") +
-#    geom_smooth(method = "lm", se = FALSE) +
-#    theme(legend.position = "none")
-#  p2 <- ggplot(evals, aes(x = age, y = score, color = ethnicity)) +
-#    geom_point() +
-#    labs(x = "Instructor age", y = "Teaching score", color = "Instructor\nEthnicity",
-#         title = "Interaction model") +
-#    geom_parallel_slopes(se = FALSE) +
-#    theme(axis.title.y = element_blank())
-#  ggsave(p1 + p2,  filename = "vignettes/evals.png", width = 8, height = 4)
+## ---- out.width = "100%", echo=FALSE, fig.align='center'----------------------
+p1 <- ggplot(evals, aes(x = age, y = score, color = ethnicity)) +
+  geom_point() +
+  labs(x = "Instructor age", y = "Teaching score", color = "Instructor\nEthnicity",
+       title = "Interaction model") +
+  geom_smooth(method = "lm", se = FALSE) +
+  theme(legend.position = "none")
+p2 <- ggplot(evals, aes(x = age, y = score, color = ethnicity)) +
+  geom_point() +
+  labs(x = "Instructor age", y = "Teaching score", color = "Instructor\nEthnicity",
+       title = "Interaction model") +
+  geom_parallel_slopes(se = FALSE) + 
+  theme(axis.title.y = element_blank())
+p1 + p2
 
-## ---- out.width = "100%", echo=FALSE, fig.align='center'-----------------
-knitr::include_graphics("evals.png")
-
-## ---- eval=FALSE---------------------------------------------------------
+## ---- eval=FALSE--------------------------------------------------------------
 #  interaction_evals <- lm(score ~ age * ethnicity, data = evals)
 #  get_regression_table(interaction_evals)
 
-## ---- echo=FALSE---------------------------------------------------------
+## ---- echo=FALSE--------------------------------------------------------------
 interaction_evals <- lm(score ~ age * ethnicity, data = evals)
 get_regression_table(interaction_evals) %>% 
   knitr::kable()
 
-## ---- eval=FALSE---------------------------------------------------------
+## ---- eval=FALSE--------------------------------------------------------------
 #  parallel_slopes_evals <- lm(score ~ age + ethnicity, data = evals)
 #  get_regression_table(parallel_slopes_evals)
 
-## ---- echo=FALSE---------------------------------------------------------
+## ---- echo=FALSE--------------------------------------------------------------
 parallel_slopes_evals <- lm(score ~ age + ethnicity, data = evals)
 get_regression_table(parallel_slopes_evals) %>% 
   knitr::kable()
 
-## ---- eval=FALSE---------------------------------------------------------
+## ---- eval=FALSE--------------------------------------------------------------
 #  ggplot(MA_schools, aes(x = perc_disadvan, y = average_sat_math, color = size)) +
 #    geom_point(alpha = 0.25) +
 #    labs(x = "Percent economically disadvantaged", y = "Math SAT Score",
@@ -192,47 +188,42 @@ get_regression_table(parallel_slopes_evals) %>%
 #         color = "School size", title = "Interaction model") +
 #    geom_smooth(method = "lm", se = FALSE)
 
-## ---- echo=FALSE, eval = FALSE-------------------------------------------
-#  # eval set to FALSE b/c patchwork does not exist on CRAN
-#  library(patchwork)
-#  p1 <- ggplot(MA_schools,
-#               aes(x = perc_disadvan, y = average_sat_math, color = size)) +
-#    geom_point(alpha = 0.25) +
-#    geom_smooth(method = "lm", se = FALSE) +
-#    labs(x = "Percent economically disadvantaged", y = "Math SAT Score",
-#         color = "School size", title = "Interaction model") +
-#    theme(legend.position = "none")
-#  p2 <- ggplot(MA_schools,
-#         aes(x = perc_disadvan, y = average_sat_math, color = size)) +
-#    geom_point(alpha = 0.25) +
-#    geom_parallel_slopes(se = FALSE) +
-#    labs(x = "Percent economically disadvantaged", y = "Math SAT Score",
-#         color = "School size", title = "Parallel slopes model")  +
-#    theme(axis.title.y = element_blank())
-#  ggsave(p1 + p2,  filename = "vignettes/MA_schools.png", width = 8, height = 4)
+## ---- out.width = "100%", echo=FALSE, fig.align='center'----------------------
+p1 <- ggplot(MA_schools, 
+             aes(x = perc_disadvan, y = average_sat_math, color = size)) +
+  geom_point(alpha = 0.25) +
+  geom_smooth(method = "lm", se = FALSE) +
+  labs(x = "Percent economically disadvantaged", y = "Math SAT Score", 
+       color = "School size", title = "Interaction model") + 
+  theme(legend.position = "none")
+p2 <- ggplot(MA_schools, 
+       aes(x = perc_disadvan, y = average_sat_math, color = size)) +
+  geom_point(alpha = 0.25) +
+  geom_parallel_slopes(se = FALSE) + 
+  labs(x = "Percent economically disadvantaged", y = "Math SAT Score", 
+       color = "School size", title = "Parallel slopes model")  +
+  theme(axis.title.y = element_blank())
+p1 + p2
 
-## ---- out.width = "100%", echo=FALSE, fig.align='center'-----------------
-knitr::include_graphics("MA_schools.png")
-
-## ---- eval=FALSE---------------------------------------------------------
+## ---- eval=FALSE--------------------------------------------------------------
 #  interaction_MA <- lm(average_sat_math ~ perc_disadvan * size, data = MA_schools)
 #  get_regression_table(interaction_MA)
 
-## ---- echo=FALSE---------------------------------------------------------
+## ---- echo=FALSE--------------------------------------------------------------
 interaction_MA <- lm(average_sat_math ~ perc_disadvan * size, data = MA_schools)
 get_regression_table(interaction_MA) %>% 
   knitr::kable()
 
-## ---- eval=FALSE---------------------------------------------------------
+## ---- eval=FALSE--------------------------------------------------------------
 #  parallel_slopes_MA <- lm(average_sat_math ~ perc_disadvan + size, data = MA_schools)
 #  get_regression_table(parallel_slopes_MA)
 
-## ---- echo=FALSE---------------------------------------------------------
+## ---- echo=FALSE--------------------------------------------------------------
 parallel_slopes_MA <- lm(average_sat_math ~ perc_disadvan + size, data = MA_schools)
 get_regression_table(parallel_slopes_MA) %>% 
   knitr::kable()
 
-## ------------------------------------------------------------------------
+## -----------------------------------------------------------------------------
 get_regression_points(score_model)
 broom::augment(score_model)