+++ R/geom_parallel_slopes.R
@@ -15,8 +15,8 @@
 #' @param n Number of points per group at which to evaluate model.
 #' @inheritParams ggplot2::layer
 #' @inheritParams ggplot2::geom_smooth
+#' @inheritParams ggplot2::stat_smooth
 #'
-#' @seealso See \code{\link[ggplot2]{stat_smooth}} for more potential arguments to the underlying stat (e.g., fullrange)
 #' @examples
 #' library(dplyr)
 #' library(ggplot2)
@@ -57,6 +57,7 @@
 geom_parallel_slopes <- function(mapping = NULL, data = NULL,
                                  position = "identity", ...,
                                  se = TRUE, formula = y ~ x, n = 100,
+                                 fullrange = FALSE, level = 0.95,
                                  na.rm = FALSE, show.legend = NA,
                                  inherit.aes = TRUE) {
   # Possibly warn about not needing to pass `method` argument
@@ -72,7 +73,11 @@ geom_parallel_slopes <- function(mapping = NULL, data = NULL,
   }
   
   # Construct layer params
-  params <- c(list(na.rm = na.rm, se = se, formula = formula, n = n), dots)
+  stat_params <- c(
+    na.rm = na.rm, se = se, formula = formula, n = n, fullrange = fullrange,
+    level = level
+  )
+  params <- c(stat_params, dots)
   
   ggplot2::layer(
     geom = ggplot2::GeomSmooth, stat = StatParallelSlopes, data = data,
@@ -86,7 +91,8 @@ StatParallelSlopes <- ggplot2::ggproto(
   
   required_aes = c("x", "y"),
   
-  compute_panel = function(data, scales, se = TRUE, formula = y ~ x, n = 100, fullrange=FALSE) {
+  compute_panel = function(data, scales, se = TRUE, formula = y ~ x, n = 100,
+                           fullrange = FALSE, level = 0.95) {
     if (nrow(data) == 0) {
       return(data[integer(0), ])
     }
@@ -99,18 +105,21 @@ StatParallelSlopes <- ggplot2::ggproto(
     # Fit model
     model <- stats::lm(formula = formula, data = data)
     
-    if (fullrange) {
-      support <- scales$x$dimension()
-    } else {
-      support <- quote(range(group_df$x, na.rm = TRUE))
-    }
-    
     # Compute prediction from model based on sequence of x-values defined for
-    # every group separately
-    stats <- lapply(
-      X = split(data, data$group), FUN = predict_per_group,
-      model = model, se = se, n = n, support=support
+    # every group
+    groups <- split(data, data$group)
+    groups_new_data <- lapply(
+      X = groups, FUN = compute_group_new_data,
+      scales = scales, n = n, fullrange = fullrange
     )
+    stats <- lapply(
+      X = groups_new_data, FUN = predict_df,
+      model = model, se = se, level = level
+    )
+    
+    # Restore columns that describe group with unique value (like color, etc.)
+    # so that they can be used in output plot
+    stats <- mapply(restore_unique_cols, stats, groups, SIMPLIFY = FALSE)
     
     # Combine predictions into one data frame
     dplyr::bind_rows(stats)
@@ -136,18 +145,25 @@ compute_model_info <- function(data, formula) {
   list(formula = formula, data = data)
 }
 
-predict_per_group <- function(group_df, model, se, n, support) {
-  # This code is a modified version of `ggplot2:::predictdf.default`
+compute_group_new_data <- function(group_df, scales, n, fullrange) {
+  if (fullrange) {
+    support <- scales$x$dimension()
+  } else {
+    support <- range(group_df$x, na.rm = TRUE)
+  }
   
-  # Create a data on which to perform prediction
-  x_support <- eval(support)
-  x_seq <- seq(x_support[1], x_support[2], length.out = n)
+  x_seq <- seq(support[1], support[2], length.out = n)
   group_seq <- rep(group_df$group[1], n)
-  new_data <- data.frame(x = x_seq, group = group_seq)
+  
+  data.frame(x = x_seq, group = group_seq)
+}
+
+predict_df <- function(model, new_data, se, level) {
+  # This code is a modified version of `ggplot2:::predictdf.default`
   
   # Perform prediction
   pred <- stats::predict(
-    model, newdata = new_data, se.fit = se,
+    model, newdata = new_data, se.fit = se, level = level,
     interval = if (se) "confidence" else "none"
   )
   
@@ -156,14 +172,10 @@ predict_per_group <- function(group_df, model, se, n, support) {
     fit <- as.data.frame(pred$fit)
     names(fit) <- c("y", "ymin", "ymax")
     
-    res <- data.frame(x = x_seq, fit, se = pred$se.fit)
+    data.frame(x = new_data$x, fit, se = pred$se.fit)
   } else {
-    res <- data.frame(x = x_seq, y = as.vector(pred))
+    data.frame(x = new_data$x, y = as.vector(pred))
   }
-  
-  # Restore columns that describe group with unique value (like color, etc.)
-  # so that they can be used in output plot
-  restore_unique_cols(res, group_df)
 }
 
 # Combination of source code from `ggplot2::StatSmooth$compute_panel` and