+++ NEWS.md
@@ -1,6 +1,6 @@
 # moderndive 0.4.0.9000
 
-* Added [Journal of Open Source Education](https://jose.theoj.org/) paper in `JOSE/`
+* Modified `vignettes/why-moderndive.Rmd` to act as submission to [Journal of Open Source Education](https://jose.theoj.org/) 
 * "Why ModernDive?" vignette now loads `patchwork` package from CRAN